pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opensre"
version = "2026.4.5"
description = "Open-source SRE agent for automated incident investigation and root cause analysis. Automatically analyzes alerts from Slack, Grafana, Datadog, and other tooling."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
  { name = "Tracer Cloud", email = "support@tracer.cloud" },
]
keywords = ["sre", "devops", "incident", "rca", "llm", "ai", "agent", "observability"]
dependencies = [
    # Core AI/Agent
    "anthropic>=0.77.1",
    "mcp>=1.27.0",
    "openai>=2.0.0",
    "langsmith>=0.7.30",
    "langgraph>=1.1.6",
    "langgraph-sdk",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    # Data validation
    "pydantic>=2.12.5,<3",
    "pydantic-settings>=2.12.0,<3",
    "kubernetes>=28.1.0",
    # HTTP and async
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "fastapi>=0.135.3",
    # Authentication
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "keyring",
    # AWS
    "boto3>=1.42.88",
    # Utilities
    "python-dotenv>=1.2.2",
    "click>=8.1.0",
    "rich>=15.0.0",
    "questionary>=2.1.1",
    "PyYAML>=6.0",
    "tzdata>=2026.1",
    # OpenTelemetry
    "opentelemetry-api>=1.41.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.41.0",
    "opentelemetry-instrumentation>=0.40b0",
    "opentelemetry-instrumentation-botocore>=0.40b0",
    "opentelemetry-instrumentation-requests>=0.40b0",
    # Tracer
    "tracer_decorator",
    # Google APIs
    "google-api-python-client>=2.194.0,<3.0.0",
    "google-auth>=2.0.0,<3.0.0",
    # MongoDB
    "pymongo>=4.6.0,<6.0.0",
    # Signature verfication for discord
    "PyNaCl>=1.5.0",
    # MariaDB
    "pymysql>=1.1.0,<2.0.0",
]

[project.urls]
Homepage = "https://github.com/Tracer-Cloud/opensre"
Repository = "https://github.com/Tracer-Cloud/opensre"
Issues = "https://github.com/Tracer-Cloud/opensre/issues"
Changelog = "https://github.com/Tracer-Cloud/opensre/releases"

[project.scripts]
opensre = "app.cli.__main__:main"
opensre-mcp = "app.entrypoints.mcp:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-xdist>=3.5.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-requests",
    "types-psycopg2",
    "types-PyMySQL",
    "types-PyYAML",
    "boto3-stubs[essential]",
]
kafka = [
    "confluent-kafka>=2.14.0",
]
clickhouse = [
    "clickhouse-connect>=0.15.1",
]
postgresql = [
    "psycopg2-binary>=2.9.0,<3.0.0",
]
azure_sql = [
    "pyodbc>=5.1.0,<6.0.0",
]

[tool.setuptools.packages.find]
include = ["app*", "tests*"]

# With pytest-xdist (-n auto), each worker writes a fragment; put them under
# .pytest_cache/ (already gitignored) instead of .coverage.<hostname>.* at repo root.
[tool.coverage.run]
data_file = ".pytest_cache/.coverage"

[tool.vulture]
paths = ["app"]
exclude = [
  "app/tools/simple_tools.py",          # explicitly skipped by the registry
  "app/dockerfile_test.py",
  "app/**/*_test.py",                   # inline pytest modules
]
ignore_decorators = [
  "@*.command",          # Click / Typer subcommands
  "@*.group",
  "@*.callback",
  "@app.*",              # FastAPI routes
  "@*router.*",          # FastAPI sub-routers (discord_router etc.)
  "@auth.on.*",          # LangGraph auth callbacks
  "@field_validator",    # Pydantic validators
  "@model_validator",
  "@validator",
  "@cache",
  "@lru_cache",
  "@tool",               # @app.tools.tool_decorator.tool registrations
  "@mcp.tool",
  "@register*",
]
ignore_names = [
  "shell_complete", "format_help", "show",        # Click overrides
  "_normalize_*",                                  # Pydantic field validators
  "_require_*",                                    # Pydantic model validators
  "on_thread_*", "on_assistant_*", "on_cron_*",    # LangGraph auth handlers
  "discord_interactions", "health_check",
  "version_check", "deep_health_check",
  "investigate_stream",
]
